+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -167,6 +167,7 @@ public class Settings extends Config {
 					EntityType.SKELETON_HORSE.name(),
 					EntityType.ZOMBIE_HORSE.name(),
 					EntityType.SHULKER.name(),
+					EntityType.ARMOR_STAND.name(),
 					EntityType.EVOKER.name(), // MC 1.11
 					EntityType.VEX.name(), // MC 1.11
 					EntityType.VINDICATOR.name(), // MC 1.11